Refactor move wear routed content to krouter package
